fix(active-child): split into .server.ts / .client.ts to prevent next/headers in client bundle
Turbopack build failed because active-child.ts imported next/headers (server-only)
but was also pulled into Client Component bundles via admin-shell.tsx, checkout-screen.tsx,
and home-client.tsx. Split into active-child.server.ts (server-only, guarded by
"server-only" package) and active-child.client.ts (pure document.cookie helpers).
Updated all six import sites to use the correct module.
