+++ apps/web/src/lib/active-child.server.ts
@@ -1,9 +1,7 @@
+import "server-only";
 import { cookies } from "next/headers";
 import { getSessionUser } from "@/lib/auth/authorization";
 import { getChildById } from "@/db/queries";
-
-const COOKIE_NAME = "uo:active-child";
-const COOKIE_MAX_AGE_SECONDS = 60 * 60 * 24 * 30; // 30 days
 
 export type ActiveChild = {
   id: string;
@@ -12,6 +10,9 @@ export type ActiveChild = {
   year: string;            // canonical short form, e.g. "9"
   rollClass: string | null;
 };
+
+const COOKIE_NAME = "uo:active-child";
+const COOKIE_MAX_AGE_SECONDS = 60 * 60 * 24 * 30; // 30 days
 
 /**
  * Server-side reader. Resolves the cookie's child UUID, ownership-checks
@@ -61,29 +62,4 @@ export async function setActiveChildCookieServer(childId: string): Promise<void>
 export async function clearActiveChildCookieServer(): Promise<void> {
   const cookieStore = await cookies();
   cookieStore.delete(COOKIE_NAME);
-}
-
-/**
- * Client-side cookie name + helpers. Used by the picker tap handler.
- * httpOnly is false on this cookie because we set it from the client.
- */
-export const ACTIVE_CHILD_COOKIE_NAME = COOKIE_NAME;
-
-export function setActiveChildCookieClient(childId: string): void {
-  if (typeof document === "undefined") return;
-  document.cookie = `${COOKIE_NAME}=${encodeURIComponent(
-    childId
-  )}; Path=/; Max-Age=${COOKIE_MAX_AGE_SECONDS}; SameSite=Lax`;
-}
-
-export function readActiveChildCookieClient(): string | null {
-  if (typeof document === "undefined") return null;
-  const escaped = COOKIE_NAME.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-  const match = document.cookie.match(new RegExp(`(?:^|; )${escaped}=([^;]*)`));
-  return match ? decodeURIComponent(match[1]) : null;
-}
-
-export function clearActiveChildCookieClient(): void {
-  if (typeof document === "undefined") return;
-  document.cookie = `${COOKIE_NAME}=; Path=/; Max-Age=0; SameSite=Lax`;
 }
